Add files via upload
1. Multiply the Number by 2

Write a function that receives a number and prints as result that number multiplied by two.

2. Student Information

You will be given 3 parameters – student name (string), age (number) and average grade (number). Your task is to print all the info about the student in the following format:

"Name: {student name}, Age: {student age}, Grade: {student grade}"

Note: The grade should be formatted to the second decimal point

3. Excellent Grade

Write a function that receives a single number and checks if the grade is excellent or not. If it is, print "Excellent", otherwise print "Not excellent".

4. Month Printer

Write a program, that takes an integer as a parameter and prints the corresponding month. If the number is more than 12 or less than 1, print "Error!"

Input

You will receive a single number.

Output

If the number is within the boundaries, print the corresponding month, otherwise print "Error!"

4. Month Printer

Write a program, that takes an integer as a parameter and prints the corresponding month. If the number is more than 12 or less than 1, print "Error!"

Input

You will receive a single number.

Output

If the number is within the boundaries, print the corresponding month, otherwise print "Error!"

6. Largest Number

Write a function that takes three number arguments as input and finds the largest of them.

Print the following text on the console: "The largest number is {number}.".

The input comes as three number arguments passed to your function.

The output should be printed to the console. 

7. Theatre Promotions

A theatre is doing a ticket sale, but they need a program to calculate the price of a single ticket. If the given age does not fit one of the categories, you should print "Error!". You can see the prices in the table below: Input

The input comes in two parameters. The first one will be the type of day (string). The second is the age of the person (number).

Output

Print the price of the ticket according to the table, or "Error!" if the age is not in the table.

Constraints

· The age will be in the interval [-1000…1000].

· The type of day will always be valid. 

8. Circle Area

Write a function that takes a single argument as input. Check the type of input argument. If it is a number, assume it is the radius of a circle and calculate the circle area. Print the area rounded to two decimal places.

If the argument type is NOT a number, print the following text on the console: "We can not calculate the circle area, because we received a {type of argument}."

The input comes as a single argument passed to your function.

The output should be printed on the console.

9. Numbers from 1 to 5

Write a function that prints all the numbers from 1 to 5 (inclusive) each on a separate line.

Hints

Create a for loop starting from 1 and continuing until 5 and print the number

10. Numbers from M to N

Write a function that receives a number M and a number N (M will always be bigger than N). Print all numbers from M to N.