Add files via upload
1. Sum First and Last Array Elements

Write a function that receives an array of numbers and prints the sum of the first and last element in that array

2. Day of Week

Write a program, which receives a number and prints the corresponding name of the day of the week (in English).

If the number is NOT a valid day, print: "Invalid day!".

3. Reverse an Array of Numbers

Write a program, which receives a number n and an array of elements. Your task is to create a new array with n numbers from the original array, reverse it and print its elements on a single line, space-separated.

4. Reverse In Place

Write a program, which receives an array of strings. Your task is to reverse the array without creating a new array. Print the resulting elements on a single line, space-separated. 

5. Phone Book

Write a function that stores information about a person's name and phone number. The input is an array of strings with space-separated name and number. Replace duplicate names. Print the result as shown.

6. Meetings

Write a function that manages meeting appointments. The input comes as an array of strings. Each string contains a weekday and person's name. For each successful meeting, print a message. If you receive the same weekday twice, the meeting cannot be scheduled so print a conflicting message. In the end, print a list of all successful meetings

7. Address Book

Write a function that stores information about a person's name and their address. The input comes as an array of strings. Each string contains the name and the address separated by a colon. If you receive the same name twice just replace the address. In the end, print the full list, sorted alphabetically by the person’s name