Word Tracker
Write a function that receives an array of words and finds occurrences of given words in that sentence.

The input will come as an array of strings. The first string will contain the words you will be looking for separated by a space. All strings after that will be the words in which you will check for a match.

Print for each word how many times it occurs. The words should be sorted by count in descending.