Info
1. Person Info

Write a function that receives 3 parameters, sets them to an object, and returns that object.

The input comes as 3 separate strings in the following order: firstName, lastName, age.

2. Convert to Object

Write a function that receives a string in JSON format and converts it to an object.

Loop through all the keys and print them with their values in format: "{key}: {value}" 

3. Convert to JSON

Write a function that receives a first name, last name, hair color and sets them to an object.

Convert the object to JSON string and print it.

Input is provided as 3 single strings in the order stated above. 

4. Cats

Write a function that receives array of strings in the following format '{cat name} {age}'.

Create a Cat class that receives in the constructor the name and the age parsed from the input.

It should also have a method named "meow" that will print "{cat name}, age {age} says Meow" on the console.

For each of the strings provided, you must create a cat object and invoke the .meow () method

5. Format Grade

Write a function that receives a grade between 2.00 and 6.00 and prints a formatted line with grade and description.

· < 3.00 - "Fail"

· >= 3.00 and < 3.50 - "Poor"

· >= 3.50 and < 4.50 - "Good"

· >= 4.50 and < 5.50 - "Very good"

· >= 5.50 - "Excellent"

6. Simple Calculator

Write a function that receives three parameters – two numbers and an operator (string) – and calculates the result depending on the operator. Operator can be 'multiply', 'divide', 'add' or 'subtract'. Try to solve this task using arrow functions.

Bonus

Solve this task without using any conditional statements (no if or switch statements or ternary operators).

Input

The input comes as parameters named numOne, numTwo, operator.