Even or Odd
You need to write unit tests for a function isOddOrEven() that checks whether the length of a passed string is even or odd.

If the passed parameter is NOT a string return undefined. If the parameter is a string return either "even" or "odd" based on the length of the string.