Food Delivery
The object should have the following functionality:

· getCategory(category) - A function that accepts one parameter: string.

o If the category is "Vegan" return the string:

"Dishes that contain no animal products."

o If the category is "Vegetarian" return the string:

"Dishes that contain no meat or fish."

o If the category is "Gluten-Free" return the string:

"Dishes that contain no gluten."

o If the category is "All" return the string:

"All available dishes."

o If the value of the string type is different from "Vegan", "Vegetarian", "Gluten-Free", "All", throw an error:

"Invalid Category!"

· addMenuItem(menuItem, maxPrice) - A function that accepts an array of objects({name: Item name, price: item price}) and number.

o You must add an element (menuItem) if the price is less or equal to maxPrice from the array to availableItems array.

o Finally, return the array length in the following string:

o "There are {availableItems.length} available menu items matching your criteria!"

o There is a need for validation for the input, as an array and

number may not always be valid. In case of submitted invalid parameters, throw an error

"Invalid Information!"

§ If passed menuItem or maxPrice parameters are not an array and number.

§ If the menuItem array has fewer than 1 item, and if maxPrice is less than 5.

· calculateOrderCost(shipping, addons, discount) - A function that accepts three parameters: array, array, and boolean.

o Calculate the total price you are going to pay depending on the chosen shipping options and addons.

§ The result must be formatted to the second digit after the decimal point.

o The available options for shipping are:

§ standard, which costs $3

§ express, which costs $5

o The available options for addons are:

§ sauce, which costs $1

§ beverage, which costs $3.5

o If the discount is true, a 15% discount should be applied. Then return the following message:

"You spend ${totalPrice} for shipping and addons with a 15% discount!"

o Else, return the following message:

o "You spend ${totalPrice} for shipping and addons!"

o You need to validate the input. If the shipping, addons, and discount are not an array, array, and boolean, throw an error:

"Invalid Information!"

o Note: Тhe totalPrice must be rounded to the second decimal